% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdmcLikelihood.R
\docType{package}
\name{ggdmcLikelihood}
\alias{ggdmcLikelihood-package}
\alias{ggdmcLikelihood}
\title{Tools for Understanding How to Fit Cognitive Hierarchical Models}
\description{
\pkg{ggdmcLikelihood} provides associated tools to access the internal data of likelihood_class
}
\author{
Yi-Shin Lin <yishinlin001@gmail.com>
}
\keyword{internal}
\keyword{package}
