% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdmcLikelihood.R
\docType{package}
\name{ggdmcLikelihood}
\alias{ggdmcLikelihood-package}
\alias{ggdmcLikelihood}
\title{Likelihood tools for the ggdmc package}
\description{
This package provides functions likelihood computations associated
with the ggdmc package. The package supports both single-subject
and multi-subject likelihood computations.
}
\author{
Yi-Shin Lin <yishinlin001@gmail.com>
}
\keyword{internal}
\keyword{package}
